+++ src/main/java/wrapper/model/constraint/Constraint.java
@@ -2,5 +2,5 @@ package wrapper.model.constraint;
 
 import lombok.NonNull;
 
-public record Constraint(int index, @NonNull ConstraintType type) {
+public record Constraint(long index, @NonNull ConstraintType type) {
 }
