+++ src/main/java/wrapper/model/variable/Variable.java
@@ -1,4 +1,4 @@
 package wrapper.model.variable;
 
-public record Variable(int index) {
+public record Variable(long index) {
 }
